Split header files into something that makes more sense
