+++ includes/gbmu.h
@@ -1,0 +1,16 @@
+/**
+ * @file gbmu.h
+ * @author Nicolas IRAGNE (nicolas.iragne@alyce.fr)
+ * @brief The main header file of the gb library.
+ * @date 2022-01-03
+ * 
+ * @copyright Copyright Alyce (c) 2022
+ */
+
+#pragma once
+
+#include "gb.h"
+#include "cpu.h"
+#include "rom.h"
+#include "memory.h"
+#include "registers.h"
