+++ includes/registers.h
@@ -1,0 +1,56 @@
+/**
+ * @file registers.h
+ * @author Nicolas IRAGNE (nicolas.iragne@alyce.fr)
+ * @brief 
+ * @date 2022-01-03
+ * 
+ * @copyright Copyright Alyce (c) 2022
+ */
+
+#include <stdint.h>
+
+struct registers_s {
+	struct {
+		union {
+			struct {
+				uint8_t f;
+				uint8_t a;
+			};
+			uint16_t af;
+		};
+	};
+	
+	struct {
+		union {
+			struct {
+			
+				uint8_t c;
+				uint8_t b;
+			};
+			uint16_t bc;
+		};
+	};
+	
+	struct {
+		union {
+			struct {
+				uint8_t e;
+				uint8_t d;
+			};
+			uint16_t de;
+		};
+	};
+	
+	struct {
+		union {
+			struct {
+				uint8_t l;
+				uint8_t h;
+			};
+			uint16_t hl;
+		};
+	};
+	
+	uint16_t	sp; // Stack pointer
+	uint16_t	pc; // Program counter
+};
